[A-**] properly broke everything with the game; complete rewrite of the rendering event system, to a dynamic one
RenderingLayerManager.py handles now RenderingLayer's. Each rendering layer defines some "environment" to be handled
in. It can be turned on and off.

It is optimised in the way that layers rendered after each other with same config are not doing anything to OpenGL

Removed all rendering events, some special ones are NOT replaced by new ones!

Signed-off-by: uuk0 <baulukas1301@googlemail.com>